refresh columns and indexes

+++ src/main/java/migrator/app/domain/table/action/TableRemoveHandler.java
@@ -2,6 +2,7 @@ package migrator.app.domain.table.action;
 
 import migrator.app.domain.table.TableContainer;
 import migrator.app.domain.table.model.Table;
+import migrator.app.migration.model.ChangeCommand;
 import migrator.lib.dispatcher.Event;
 import migrator.lib.dispatcher.EventHandler;
 
@@ -18,6 +19,10 @@ public class TableRemoveHandler implements EventHandler {
             return;
         }
         Table table = (Table) event.getValue();
+        if (!table.getChangeCommand().isType(ChangeCommand.CREATE)) {
+            table.getChangeCommand().setType(ChangeCommand.DELETE);
+            return;
+        }
         this.tableContainer.tableStore().remove(table);
         this.tableContainer.tableRepository().removeWith(table);
     }
